requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
numpy==1.24.3
scikit-learn==1.3.0
opencv-python-headless==4.8.1.78
Pillow==10.1.0
python-multipart==0.0.6
websockets==12.0